+++ src/module/AnnotationProvider.js
@@ -1,6 +1,5 @@
 ﻿import React, { createContext, useState, useContext } from "react";
-
-import { fabric } from "fabric"
+import { useEffect } from "react/cjs/react.development";
 
 const AnnotationContext = createContext();
 export const useAnnotation = () => useContext(AnnotationContext);
@@ -10,7 +9,8 @@ export default function AnnotationProvider({ children }) {
   const [labelsData, setLabelsData] = useState(null);
   const [oldLabels, setOldLabels] = useState();
   const [newLables, setNewLabels] = useState();
-  const [Canvas, setCanvas] = useState();  
+  const [isDrawingActive, setIsDrawingActive] = useState(false);  
+
   /*
   const changeAnnotationScene = (cnt) =>
     setannotationScene(cnt);
@@ -24,7 +24,7 @@ export default function AnnotationProvider({ children }) {
       setLabelsData(data);
 
   return (
-    <AnnotationContext.Provider value={{ labelsData, changeLabelsData, oldLabels, setOldLabels, newLables, setNewLabels, Canvas, setCanvas }}>
+    <AnnotationContext.Provider value={{ labelsData, changeLabelsData, oldLabels, setOldLabels, newLables, setNewLabels, isDrawingActive, setIsDrawingActive }}>
       {children}
     </AnnotationContext.Provider>
   )
