+++ src/style.css
@@ -1092,7 +1092,6 @@ video {
 
 
 /* ローディングアニメーション */
-
 .spinner {
   position: absolute;
   top: 50%;
@@ -1308,4 +1307,52 @@ details[open] summary {
   position: absolute;
   top: 0;
   left: 0;
+}
+
+/* api読み込み中アニメーション */
+.loader,
+.loader:after {
+  border-radius: 50%;
+  width: 10em;
+  height: 10em;
+}
+.loader {
+  position: absolute;
+  top: 50%;
+  left: 47%;
+  transform: translate(-50%, -50%);
+  
+  text-align: center;
+  font-size: 10px;
+  /* position: relative; */
+  text-indent: -9999em;
+  border-top: 1.1em solid rgba(15,15,15, 0.2);
+  border-right: 1.1em solid rgba(15,15,15, 0.2);
+  border-bottom: 1.1em solid rgba(15,15,15, 0.2);
+  border-left: 1.1em solid #0f0f0f;
+  -webkit-transform: translateZ(0);
+  -ms-transform: translateZ(0);
+  transform: translateZ(0);
+  -webkit-animation: load8 1.1s infinite linear;
+  animation: load8 1.1s infinite linear;
+}
+@-webkit-keyframes load8 {
+  0% {
+    -webkit-transform: rotate(0deg);
+    transform: rotate(0deg);
+  }
+  100% {
+    -webkit-transform: rotate(360deg);
+    transform: rotate(360deg);
+  }
+}
+@keyframes load8 {
+  0% {
+    -webkit-transform: rotate(0deg);
+    transform: rotate(0deg);
+  }
+  100% {
+    -webkit-transform: rotate(360deg);
+    transform: rotate(360deg);
+  }
 }
